Split the code into two parts in order to compile on WASM.
