+++ linera-ethereum/src/lib.rs
@@ -11,5 +11,9 @@
 pub mod client;
 pub mod common;
 
+#[cfg(not(target_arch = "wasm32"))]
+pub mod provider;
+
 /// Helper types for tests and similar purposes.
+#[cfg(not(target_arch = "wasm32"))]
 pub mod test_utils;
