+++ linera-ethereum/src/test_utils/mod.rs
@@ -17,7 +17,8 @@ use alloy_primitives::Address;
 use linera_storage_service::child::get_free_port;
 use url::Url;
 
-use crate::client::{EthereumClient, EthereumClientSimplified, EthereumQueries, HttpProvider};
+use crate::provider::{EthereumClient, HttpProvider};
+use crate::client::{EthereumClientSimplified, EthereumQueries};
 
 sol!(
     #[allow(missing_docs)]
